requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.108.0
greenlet==3.2.0
h11==0.14.0
idna==3.10
psycopg2==2.9.8
psycopg2-binary==2.9.8
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.0.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.20
starlette==0.32.0.post1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.25.0
